# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is the API documentation for my Laravel application.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/location": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Retrieve a list of all locations",
                "operationId": "779a41c3b7bb0377e94b79a2a74aa408",
                "responses": {
                    "200": {
                        "description": "Location list retrieved successfully"
                    },
                    "404": {
                        "description": "Location list not found"
                    }
                }
            }
        },
        "/api/v1/location_register": {
            "post": {
                "tags": [
                    "Location"
                ],
                "summary": "Store a new location",
                "operationId": "7855b7307ed44a5e61d37d50830b71b6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Location created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/location_update/{id}": {
            "put": {
                "tags": [
                    "Location"
                ],
                "summary": "Update a location",
                "operationId": "962259642b96e992182647925b2258fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Location updated successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            }
        },
        "/api/v1/location_destroy/{id}": {
            "delete": {
                "tags": [
                    "Location"
                ],
                "summary": "Delete a location",
                "operationId": "70917ac4de6c32d82c1cb1bba78817c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location deleted successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            }
        },
        "/api/v1/location_show/{id}": {
            "get": {
                "tags": [
                    "Location"
                ],
                "summary": "Retrieve a specific location",
                "operationId": "551c1e347b136d96d2a4400da49fbd13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the location to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Location retrieved successfully"
                    },
                    "404": {
                        "description": "Location not found"
                    }
                }
            }
        },
...
    },
...
}